+++ src/idle-sdk.core/Scene/SceneLayer.cs
@@ -1,0 +1,9 @@
+namespace IdleSdk.Core.Scene;
+
+public enum SceneLayer
+{
+    Background,
+    Midground,
+    Foreground,
+    Overlay
+}